+++ lib/types/versioning.ts
@@ -1,0 +1,8 @@
+export type RangeStrategy =
+  | 'auto'
+  | 'bump'
+  | 'future'
+  | 'pin'
+  | 'replace'
+  | 'update-lockfile'
+  | 'widen';
